\name{agilent.miRNA.cbind}
\alias{agilent.miRNA.cbind}
\title{cbind Agilent miRNA datasets
Combine 2 Agilent miRNA objects side by side. Since Agilent objects are
lists of tables, this works on all relevant elements}
\usage{
  agilent.miRNA.cbind(x, y)
}
\arguments{
  \item{x}{an Agilent miRNA object (which for now is a
  \code{list})}

  \item{y}{an Agilent miRNA object (which for now is a
  \code{list})}
}
\value{
  a new, larger Agilent miRNA object
}
\description{
  cbind Agilent miRNA datasets Combine 2 Agilent miRNA
  objects side by side. Since Agilent objects are lists of
  tables, this works on all relevant elements
}
\author{
  Mark Cowley, 2011-08-02
}

