\name{as.data.frame.AnnotatedDataFrame}
\alias{as.data.frame.AnnotatedDataFrame}
\title{Coerce to a Data Frame}
\usage{
  \method{as.data.frame}{AnnotatedDataFrame} (x, ...)
}
\arguments{
  \item{x}{an AnnotatedDataFrame}

  \item{\dots}{unused}
}
\value{
  a \code{data.frame}
}
\description{
  Coerce to a Data Frame
}
\author{
  Mark Cowley, 2012-07-24
}

