\name{import.Agilent.GeneView.File}
\alias{import.Agilent.GeneView.File}
\title{Import Agilent GeneView files}
\usage{
  import.Agilent.GeneView.File(file)
}
\arguments{
  \item{file}{the path to the file to import}
}
\description{
  Import a GeneView file produced by Agilent Feature
  Extractor.
}
\author{
  Mark Cowley, 2008-08-01
}

