\name{import.agilent}
\alias{import.agilent}
\title{Import an Agilent result file}
\usage{
  import.agilent(f)
}
\arguments{
  \item{f}{the path to an Agilent Microarray TXT file}
}
\description{
  Import a txt file produced by the Agilent Feature
  Extraction program. More specifically, import the FEATURE
  rows, which constitute the 3rd section of the file, which
  is the majority.
}
\author{
  Mark Cowley, 2008-08-29
}

