\name{is.topTable.list}
\alias{is.topTable.list}
\title{Is the argument a single topTable, or a list of topTables}
\usage{
  is.topTable.list(tt)
}
\arguments{
  \item{tt}{a \code{data.frame}, or list of
  \code{data.frame}'s}
}
\value{
  logical: \code{TRUE} if \code{tt} is a list of
  \code{topTable}'s, \code{FALSE} otherwise
}
\description{
  Is the argument a single topTable, or a list of topTables
}
\author{
  Mark Cowley, 2009-01-28
}

