\docType{package}
\name{microarrays-package}
\alias{microarrays}
\alias{microarrays-package}
\title{microarrays package title}
\description{
  microarrays package description
}
\details{
  \tabular{ll}{ Package: \tab microarrays\cr Type: \tab
  Package\cr Version: \tab
  \Sexpr[stage=build]{packageDescription("microarrays")$Version}\cr
  Date: \tab \Sexpr[stage=build]{format(Sys.time(),
  "\%Y-\%m-\%d")}\cr License: \tab GPL-2\cr LazyLoad: \tab
  yes\cr }
}
\author{
  Mark Cowley Maintainer: Mark Cowley
  \email{m.cowley@garvan.org.au}
}
\keyword{package}

