\name{round.topTable}
\alias{round.topTable}
\title{Round the numeric columns in a topTable
Take a top table, and round the columns to \code{digits} places}
\usage{
  \method{round}{topTable} (x, digits = 4)
}
\arguments{
  \item{x}{a topTable object}

  \item{digits}{the number of digits to round to}
}
\value{
  a topTable object with numerical columns rounded.
}
\description{
  Round the numeric columns in a topTable Take a top table,
  and round the columns to \code{digits} places
}
\author{
  Mark Cowley
}

