\docType{methods}
\name{sort}
\alias{sort}
\alias{sort,data.frame-method}
\alias{sort,ExpressionSet-method}
\alias{sort,LumiBatch-method}
\alias{sort,matrix-method}
\title{sort}
\arguments{
  \item{FUN}{a sort function. if \code{\link{x}} is 2D or
  more, then this is applied to the 1st dimension (ie the
  rows)}

  \item{FUN}{a sort function. if \code{\link{x}} is 2D or
  more, then this is applied to the 1st dimension (ie the
  rows)}
}
\value{
  something

  something
}
\description{
  sort

  sort
}
\author{
  Mark Cowley

  Mark Cowley
}

