\docType{package}
\name{microarrays-package}
\alias{microarrays}
\alias{microarrays-package}
\title{mjc's microarrays package}
\description{
  Collection of code to facilitate microarray analysis.
  This includes S4 methods to make dealing with Illumina
  data easier, collapsing and subsetting ExpressionSet's,
  averaging replicates, lowess, plotting etc\dots
}
\details{
  \tabular{ll}{ Package: \tab microarrays\cr Type: \tab
  Package\cr Version: \tab 1.0.1\cr Date: \tab
  2012-07-31\cr License: \tab GPL-2\cr LazyLoad: \tab
  yes\cr }
}
\author{
  Mark Cowley Maintainer: Mark Cowley
  \email{m.cowley@garvan.org.au}
}
\keyword{package}

